+++ app/ui/up/header/nav-bar.tsx
@@ -12,7 +12,7 @@ import {
   MagnifyingGlassIcon,
 } from "@heroicons/react/20/solid";
 import HeaderTag from "./header-tag";
-import Dropdown from "@/app/ui/up/dropdown-header";
+import Dropdown from "@/app/ui/up/header/dropdown-header";
 
 // text area to be toggled appear or not in certain pages
 export default function NavBar() {
